Modularize app layout and debugger
